feat: Se agrego Estanque y Producto al Repositorio

+++ Backend/src/main/java/com/graspymar/ims/dto/ProductoRequestDTO.java
@@ -23,7 +23,7 @@ public class ProductoRequestDTO {
     private String unidadMedida;
 
     @NotNull
-    private Double stockActual;
+    private BigDecimal stockActual;
 
     @NotNull
     private BigDecimal precioVenta;
