+++ Backend/src/main/java/com/graspymar/ims/dto/ProductoResponseDTO.java
@@ -18,7 +18,7 @@ public class ProductoResponseDTO {
 
     private String unidadMedida;
 
-    private Double stockActual;
+    private BigDecimal stockActual;
 
     private BigDecimal precioVenta;
 }
